+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 Released
 --------
 
+0.5.0 - 28 October 2023
+===================
+**Breaking Changes**
+- Interface changed of Listable traig
+**Bugfix**
+- Selected widgets height is correctly calculated
+
+The api should be more stable from now on.
+
 0.4.0 - 25 October 2023
 ===================
 **Breaking Changes**